+++ crates/zizmor/src/audit/bot_conditions.rs
@@ -1,15 +1,19 @@
 use std::{ops::Deref, sync::LazyLock};
 
 use github_actions_expressions::{
-    BinOp, Expr, UnOp,
+    BinOp, Expr, SpannedExpr, UnOp,
     context::{Context, ContextPattern},
 };
-use github_actions_models::common::{If, expr::ExplicitExpr};
+use github_actions_models::common::If;
 
 use super::{Audit, AuditLoadError, AuditState, audit_meta};
 use crate::{
-    finding::{Confidence, Severity, location::Locatable as _},
+    finding::{
+        Confidence, Severity,
+        location::{Locatable as _, Subfeature},
+    },
     models::workflow::JobExt,
+    utils::ExtractedExpr,
 };
 
 pub(crate) struct BotConditions;
@@ -88,13 +92,18 @@ impl Audit for BotConditions {
         }
 
         for (expr, parent, if_loc) in conds {
-            if let Some(confidence) = Self::bot_condition(expr) {
+            if let Some((subfeature, confidence)) = Self::bot_condition(expr) {
                 findings.push(
                     Self::finding()
                         .severity(Severity::High)
                         .confidence(confidence)
                         .add_location(parent)
-                        .add_location(if_loc.primary().annotated("actor context may be spoofable"))
+                        .add_location(
+                            if_loc
+                                .primary()
+                                .subfeature(subfeature)
+                                .annotated("actor context may be spoofable"),
+                        )
                         .build(job.parent())?,
                 );
             }
@@ -105,8 +114,11 @@ impl Audit for BotConditions {
 }
 
 impl BotConditions {
-    fn walk_tree_for_bot_condition(expr: &Expr, dominating: bool) -> (bool, bool) {
-        match expr {
+    fn walk_tree_for_bot_condition<'a, 'src>(
+        expr: &'a SpannedExpr<'src>,
+        dominating: bool,
+    ) -> (Option<&'a SpannedExpr<'src>>, bool) {
+        match expr.deref() {
             // We can't easily analyze the call's semantics, but we can
             // check to see if any of the call's arguments are
             // bot conditions. We treat a call as non-dominating always.
@@ -118,8 +130,8 @@ impl BotConditions {
             | Expr::Context(Context { parts: exprs, .. }) => exprs
                 .iter()
                 .map(|arg| Self::walk_tree_for_bot_condition(arg, false))
-                .reduce(|(bc, _), (bc_next, _)| (bc || bc_next, false))
-                .unwrap_or((false, dominating)),
+                .reduce(|(bc, _), (bc_next, _)| (bc.or(bc_next), false))
+                .unwrap_or((None, dominating)),
             Expr::Index(expr) => Self::walk_tree_for_bot_condition(expr, dominating),
             Expr::BinOp { lhs, op, rhs } => match op {
                 // || is dominating.
@@ -127,7 +139,7 @@ impl BotConditions {
                     let (bc_lhs, _) = Self::walk_tree_for_bot_condition(lhs, true);
                     let (bc_rhs, _) = Self::walk_tree_for_bot_condition(rhs, true);
 
-                    (bc_lhs || bc_rhs, true)
+                    (bc_lhs.or(bc_rhs), true)
                 }
                 // == is trivially dominating.
                 BinOp::Eq => match (lhs.as_ref().deref(), rhs.as_ref().deref()) {
@@ -138,16 +150,16 @@ impl BotConditions {
                             || (SPOOFABLE_ACTOR_ID_CONTEXTS.iter().any(|x| x.matches(ctx))
                                 && BOT_ACTOR_IDS.contains(&lit.as_str().as_ref()))
                         {
-                            (true, true)
+                            ((Some(expr)), true)
                         } else {
-                            (false, true)
+                            (None, true)
                         }
                     }
-                    (lhs, rhs) => {
+                    (_, _) => {
                         let (bc_lhs, _) = Self::walk_tree_for_bot_condition(lhs, true);
                         let (bc_rhs, _) = Self::walk_tree_for_bot_condition(rhs, true);
 
-                        (bc_lhs || bc_rhs, true)
+                        (bc_lhs.or(bc_rhs), true)
                     }
                 },
                 // Every other binop is non-dominating.
@@ -155,7 +167,7 @@ impl BotConditions {
                     let (bc_lhs, _) = Self::walk_tree_for_bot_condition(lhs, false);
                     let (bc_rhs, _) = Self::walk_tree_for_bot_condition(rhs, false);
 
-                    (bc_lhs || bc_rhs, false)
+                    (bc_lhs.or(bc_rhs), false)
                 }
             },
             Expr::UnOp { op, expr } => match op {
@@ -170,18 +182,14 @@ impl BotConditions {
                 // an explicitly toggled condition, and `None` for no condition.
                 UnOp::Not => Self::walk_tree_for_bot_condition(expr, false),
             },
-            _ => (false, dominating),
+            _ => (None, dominating),
         }
     }
 
-    fn bot_condition(expr: &str) -> Option<Confidence> {
-        // TODO: Remove clones here.
-        let bare = match ExplicitExpr::from_curly(expr) {
-            Some(raw_expr) => raw_expr.as_bare().to_string(),
-            None => expr.to_string(),
-        };
+    fn bot_condition<'doc>(expr: &'doc str) -> Option<(Subfeature<'doc>, Confidence)> {
+        let unparsed = ExtractedExpr::new(expr);
 
-        let Ok(expr) = Expr::parse(&bare) else {
+        let Ok(expr) = Expr::parse(unparsed.as_bare()) else {
             tracing::warn!("couldn't parse expression: {expr}");
             return None;
         };
@@ -193,9 +201,11 @@ impl BotConditions {
         // always passes if the actor is dependabot[bot].
         match Self::walk_tree_for_bot_condition(&expr, true) {
             // We have a bot condition and it dominates the expression.
-            (true, true) => Some(Confidence::High),
+            (Some(expr), true) => Some((Subfeature::from_spanned_expr(expr, 0), Confidence::High)),
             // We have a bot condition but it doesn't dominate the expression.
-            (true, false) => Some(Confidence::Medium),
+            (Some(expr), false) => {
+                Some((Subfeature::from_spanned_expr(expr, 0), Confidence::Medium))
+            }
             // No bot condition.
             (..) => None,
         }
@@ -245,7 +255,7 @@ mod tests {
                 Confidence::Medium,
             ),
         ] {
-            assert_eq!(BotConditions::bot_condition(cond).unwrap(), *confidence);
+            assert_eq!(BotConditions::bot_condition(cond).unwrap().1, *confidence);
         }
     }
 }