+++ crates/zizmor/src/finding/location.rs
@@ -2,7 +2,8 @@
 
 use std::{ops::Range, sync::LazyLock};
 
-use crate::{audit::AuditInput, models::AsDocument as _, registry::InputKey};
+use crate::{audit::AuditInput, models::AsDocument, registry::InputKey};
+use github_actions_expressions::SpannedExpr;
 use line_index::{LineCol, TextSize};
 use regex::Regex;
 use serde::Serialize;
@@ -107,6 +108,23 @@ macro_rules! route {
     };
 }
 
+/// Represents a "subfeature" of a symbolic location, such as a substring
+/// within a YAML string.
+#[derive(Serialize, Clone, Debug)]
+pub(crate) struct Subfeature<'doc> {
+    pub(crate) after: usize,
+    pub(crate) fragment: &'doc str,
+}
+
+impl<'doc> Subfeature<'doc> {
+    pub(crate) fn from_spanned_expr(expr: &SpannedExpr<'doc>, bias: usize) -> Subfeature<'doc> {
+        Self {
+            after: expr.span.start + bias,
+            fragment: expr.raw,
+        }
+    }
+}
+
 /// Represents a symbolic location.
 #[derive(Serialize, Clone, Debug)]
 pub(crate) struct SymbolicLocation<'doc> {
@@ -125,6 +143,9 @@ pub(crate) struct SymbolicLocation<'doc> {
     /// A symbolic route (of keys and indices) to the final location.
     pub(crate) route: Route<'doc>,
 
+    /// An optional subfeature for the symbolic location.
+    pub(crate) subfeature: Option<Subfeature<'doc>>,
+
     /// The kind of location.
     pub(crate) kind: LocationKind,
 }
@@ -136,8 +157,15 @@ impl<'doc> SymbolicLocation<'doc> {
             annotation: self.annotation.clone(),
             link: None,
             route: self.route.with_keys(keys),
+            subfeature: None,
             kind: self.kind,
         }
+    }
+
+    /// Adds a subfeature to the current `SymbolicLocation`.
+    pub(crate) fn subfeature(mut self, subfeature: Subfeature<'doc>) -> SymbolicLocation<'doc> {
+        self.subfeature = Some(subfeature);
+        self
     }
 
     /// Adds a human-readable annotation to the current `SymbolicLocation`.
@@ -177,30 +205,66 @@ impl<'doc> SymbolicLocation<'doc> {
         self,
         document: &'doc yamlpath::Document,
     ) -> anyhow::Result<Location<'doc>> {
-        // If we don't have a path into the workflow, all
-        // we have is the workflow itself.
-        let feature = if self.route.components.is_empty() {
-            document.root()
-        } else {
-            let mut builder = yamlpath::QueryBuilder::new();
+        let (extracted, location, feature) = match &self.subfeature {
+            Some(subfeature) => {
+                // If we have a subfeature, we have to extract its exact
+                // parent feature.
+                let feature = match self.route.to_query() {
+                    Some(query) => document.query_exact(&query)?.ok_or_else(|| {
+                        // This should never fail in practice, unless our
+                        // route is malformed or ends in a key-only feature
+                        // (e.g. `foo:`). The latter shouldn't really happen,
+                        // since there's no meaningful subfeature in that case.
+                        anyhow::anyhow!(
+                            "failed to extract exact feature for symbolic location: {}",
+                            self.annotation
+                        )
+                    })?,
+                    None => document.root(),
+                };
 
-            for component in &self.route.components {
-                builder = match component {
-                    RouteComponent::Key(key) => builder.key(key),
-                    RouteComponent::Index(idx) => builder.index(*idx),
-                }
+                let extracted = document.extract(&feature);
+
+                let subfeature_span = {
+                    let bias = feature.location.byte_span.0;
+                    let start = &extracted[subfeature.after..]
+                        .find(subfeature.fragment)
+                        .ok_or_else(|| {
+                            anyhow::anyhow!(
+                                "failed to find subfeature '{}' in feature '{}'",
+                                subfeature.fragment,
+                                extracted
+                            )
+                        })?;
+
+                    (start + bias)..(start + bias + subfeature.fragment.len())
+                };
+
+                (
+                    extracted,
+                    ConcreteLocation::from_span(subfeature_span, document),
+                    feature,
+                )
             }
+            None => {
+                let feature = match self.route.to_query() {
+                    Some(query) => document.query_pretty(&query)?,
+                    None => document.root(),
+                };
 
-            let query = builder.build();
-
-            document.query_pretty(&query)?
+                (
+                    document.extract_with_leading_whitespace(&feature),
+                    ConcreteLocation::from(&feature.location),
+                    feature,
+                )
+            }
         };
 
         Ok(Location {
             symbolic: self,
             concrete: Feature {
-                location: ConcreteLocation::from(&feature.location),
-                feature: document.extract_with_leading_whitespace(&feature),
+                location,
+                feature: extracted,
                 comments: document
                     .feature_comments(&feature)
                     .into_iter()
@@ -269,6 +333,20 @@ impl ConcreteLocation {
             start_point,
             end_point,
             offset_span,
+        }
+    }
+
+    pub(crate) fn from_span(span: Range<usize>, doc: &yamlpath::Document) -> Self {
+        let start = TextSize::new(span.start as u32);
+        let end = TextSize::new(span.end as u32);
+
+        let start_point = doc.line_index().line_col(start);
+        let end_point = doc.line_index().line_col(end);
+
+        Self {
+            start_point: start_point.into(),
+            end_point: end_point.into(),
+            offset_span: span.clone(),
         }
     }
 }